requirements.txt
alembic==0.9.5
amqp==2.2.2
APScheduler==2.1.2
asn1crypto==0.23.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.1
chardet==3.0.4
click==6.7
cryptography==2.0.3
enum34==1.1.6
evernote==1.25.0
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.3.0
funcsigs==1.0.2
futures==3.1.1
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
M2Crypto==0.26.4
Mako==1.0.7
MarkupSafe==1.0
oauth2==1.9.0.post1
passlib==1.7.1
pycparser==2.18
pycrypto==2.6.1
PyMySQL==0.7.11
pyOpenSSL==17.3.0
python-dateutil==2.6.1
python-editor==1.0.3
python-engineio==1.7.0
python-socketio==1.8.1
pytz==2017.2
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
typing==3.6.2
tzlocal==1.4
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2
